refactor: use go-ci-env
