+++ go.mod
@@ -6,6 +6,7 @@ require (
 	github.com/antonmedv/expr v1.8.8
 	github.com/kr/pretty v0.1.0 // indirect
 	github.com/stretchr/testify v1.6.1
+	github.com/suzuki-shunsuke/go-ci-env v0.1.0
 	github.com/suzuki-shunsuke/go-error-with-exit-code v1.0.0
 	github.com/suzuki-shunsuke/go-httpclient v1.0.0
 	github.com/suzuki-shunsuke/go-timeout v1.0.0
