+++ pkg/cmd/post.go
@@ -2,7 +2,6 @@ package cmd
 
 import (
 	"errors"
-	"io"
 	"os"
 	"strings"
 
@@ -112,9 +111,10 @@ func (runner Runner) postAction(c *cli.Context) error {
 		return err
 	}
 
-	pt := platform.Get(os.Getenv, func(p string) (io.ReadCloser, error) {
-		return os.Open(p)
-	})
+	var pt api.Platform
+	if p, f := platform.Get(); f {
+		pt = p
+	}
 
 	var cmt api.Commenter
 	if opts.DryRun {
